1-strdup.c

+++ 0x0B-malloc_free/1-strdup.c
@@ -11,7 +11,7 @@
 char *_strdup(char *str)
 {
 	int i = 0;
-	size_t size = sizeof(str);
+	int  size = sizeof(str);
 	char *n = (char *)malloc(size);
 		if (n == NULL)
 		{
